improvements

+++ customer/views.py
@@ -9,7 +9,11 @@ from request.models import Xpref
 from request.models import PrefSpec
 from request.models import Payment
 from request import views2
+from django.contrib import messages
+from customer import forms
 
+
+import request.functions as funcs
 from django.views import View
 
 class TestView(View):
@@ -64,36 +68,46 @@ def customer_read(request):
                 for payment in payments:
                     print('             payment: ' + str(payment.amount))
                 x += 1
-
-
-
     print('total orders ' + str(x))
 
     return render(request, 'customer/read.html')
 
+
 @login_required
 def customer_form(request):
+    can_add = funcs.has_perm_or_is_owner(request.user, 'customer.add_customer')
+    if not can_add:
+        messages.error(request, 'Sorry, No access for you')
+        return redirect('errorpage')
     customer_types = Type.objects.all()
+    customer_form = forms.CustomerForm()
     return render(request, 'customer/customer.html', {
-        'customer_types': customer_types
+        'customer_types': customer_types,
+        'c_form': customer_form
     })
+
 
 @login_required
 def customer_insert(request):
+    can_add = funcs.has_perm_or_is_owner(request.user, 'customer.add_customer')
+    if not can_add:
+        messages.error(request, 'Sorry, No access for you')
+        return redirect('errorpage')
     all_customers = Customer.objects.all()
     customer_types = Type.objects.all()
     customer_type = Type.objects.get(pk=request.POST['type'])
-    customer = Customer()
-    customer.name = request.POST['name']
-    customer.code = request.POST['code']
-    customer.type = customer_type
-    customer.save()
+    customer_to_insert = Customer()
+    customer_to_insert.name = request.POST['name']
+    customer_to_insert.code = request.POST['code']
+    customer_to_insert.type = customer_type
+    customer_to_insert.save()
     msg = 'Customer added successfully'
     return render(request, 'customer/index.html', {
         'msg': msg,
         'customer_types': customer_types,
         'customers': all_customers
     })
+
 
 @login_required
 def customer_index(request):
@@ -127,10 +141,19 @@ def customer_read2(request, customer_pk):
 
 @login_required
 def customer_edit(request, customer_pk):
+    customer_instance = Customer.objects.get(pk=customer_pk)
+    can_edit = funcs.has_perm_or_is_owner(request.user, 'customer.edit_customer')
+    if not can_edit:
+        messages.error(request, 'Sorry, No access for you')
+        return redirect('errorpage')
     pass
 
 
 def customer_delete(request, customer_pk):
+    can_delete = funcs.has_perm_or_is_owner(request.user, 'customer.delete_customer')
+    if not can_delete:
+        messages.error(request, 'Sorry, No access for you')
+        return redirect('errorpage')
     customer = Customer.objects.get(pk=customer_pk)
     customer_name = customer.name
     customer.delete()
@@ -138,29 +161,52 @@ def customer_delete(request, customer_pk):
     return redirect('customer_index')
 
 
-
-
 # Type functions:
 @login_required
 def type_form(request):
+    can_add = funcs.has_perm_or_is_owner(request.user, 'customer.add_type')
+    if not can_add:
+        messages.error(request, 'No access')
+        return redirect('errorpage')
     pass
+
 
 @login_required
 def type_insert(request):
+    can_add = funcs.has_perm_or_is_owner(request.user, 'customer.add_type')
+    if not can_add:
+        messages.error(request, 'No access')
+        return redirect('errorpage')
+
     pass
+
 
 @login_required
 def type_index(request):
     pass
 
+
 @login_required
 def type_read(request, type_pk):
     pass
 
+
 @login_required
 def type_edit(request, type_pk):
+    can_edit = funcs.has_perm_or_is_owner(request.user, 'customer.edit_type')
+    if not can_edit:
+        messages.error(request, 'No access')
+        return redirect('errorpage')
+
     pass
+
 
 @login_required
 def type_delete(request, type_pk):
+    can_delete = funcs.has_perm_or_is_owner(request.user, 'customer.delete_type')
+    if not can_delete:
+        messages.error(request, 'No access')
+        return redirect('errorpage')
+
     pass
+
